requirements.txt
Django==1.10.8
django-crispy-forms==1.7.0
django-haystack==2.6.1
djangorestframework==3.7.7
elasticsearch==6.1.1
pytz==2017.3
urllib3==1.22
